#339 add new cython_test folder

+++ cython_test/example_cython.pyx
@@ -1,5 +1,5 @@
 def test(int x):
-    cdef int y = 0
+    cdef double y = 0
     cdef int i
     for i in range(x):
         y += i
